Re added Build.CS file
I'm stupid and I realized that it was causing errors due to a duplicate file in the tmp folder outside the repo but in the Source folder of my projects.
Apparently UE4 doesn't differentiate between folders named after the project and folders that aren't. It just looks for Build.CS files.